requirements.txt
# L-CRP requirements
# python3.8 environment
zennit-crp==0.6.0
torch==1.13.1
torchvision==0.14.1
albumentations==1.3.0
opencv-python==4.7.0.72
segmentation_models_pytorch==0.3.2
numpy==1.24.4
scipy>=1.4.1
tqdm>=4.41.0
addict>=2.4.0
thop  # FLOPs computation
yacs
natsort
# PCX requirements
h5py>=3.11.0
# app.py requirements
flask>=2.2.5
requests==2.26.0
Werkzeug>=3.0.6
Jinja2==3.1.3
minio==7.2.10
pillow>=8.0.0
tifffile>=2021.7.2
redis>=4.0.0
rq>=1.10.0
supervisor>=4.2.0
plotly 
